src/main/java/br/com/caelum/leilao/servico/Avaliador.java
package br.com.caelum.leilao.servico;

import br.com.caelum.leilao.dominio.Lance;
import br.com.caelum.leilao.dominio.Leilao;

import java.util.*;


public class Avaliador {

    private double maiorDeTodos = Double.NEGATIVE_INFINITY;
    private double menorDeTodos = Double.POSITIVE_INFINITY;
    private List<Lance> maiores;

    public void avalia(Leilao leilao){
        for (Lance lance : leilao.getLances()) {
            if(lance.getValor() > maiorDeTodos) {
                maiorDeTodos = lance.getValor();
            }if (lance.getValor() < menorDeTodos) {
                menorDeTodos = lance.getValor();
            }
        }

        maiores = new ArrayList<Lance>(leilao.getLances());
        Collections.sort(maiores, new Comparator<Lance>() {

            public int compare(Lance o1, Lance o2) {
                if(o1.getValor() < o2.getValor()) return 1;
                if(o1.getValor() > o2.getValor()) return -1;
                return 0;
            }
        });
        pegaOsMaioresNo(leilao);
    }

    public double getMaiorValor() {
        return maiorDeTodos;
    }

    public double getMenorValor() {
        return menorDeTodos;
    }

    public List<Lance> getTresMaiores() {
        return this.maiores;
    }

    public Double media(List<Lance> lances) {
        Double total = 0.0;
        for(Lance lance: lances){
            total += lance.getValor();
        }

        Double media = total/lances.size();
        return media;
    }

    private void pegaOsMaioresNo(Leilao leilao) {
        maiores = new ArrayList<Lance>(leilao.getLances());
        Collections.sort(maiores, new Comparator<Lance>() {
            public int compare(Lance o1, Lance o2) {
                if(o1.getValor() < o2.getValor()) return 1;
                if(o1.getValor() > o2.getValor()) return -1;
                return 0;
            }
        });
        maiores = maiores.subList(0, maiores.size() > 3 ? 3 : maiores.size());
    }
}
